+++ packages/ui-scripts/lib/commands/tag.js
@@ -29,11 +29,11 @@ import {
   info,
   confirm
 } from '@instructure/command-utils'
-import { createNPMRCFile } from '../utils/npm.js'
+import { createNPMRCFile, cleanupNPMRCFile } from '../utils/npm.js'
 
 export default {
   command: 'tag',
-  desc: 'Add/remove/list npm distribution tag',
+  desc: 'Add/remove/list pnpm distribution tag',
   builder: (yargs) => {
     yargs.option('command', {
       type: 'string',
@@ -58,54 +58,59 @@ export default {
 
 async function distTag(command, versionToTag, tag) {
   createNPMRCFile()
-  if (command === 'add') {
-    info(
-      `Command: "npm dist-tag ${command}". Version to tag as "${tag}": ${versionToTag}`
-    )
-  } else if (command === 'rm') {
-    info(`Command: "npm dist-tag ${command} ${tag}".`)
-  } else if (command === 'ls') {
-    info(`Command: "npm dist-tag ${command}".`)
-  }
-  const reply = await confirm('Continue? [y/n]\n')
-  if (!['Y', 'y'].includes(reply.trim())) {
-    process.exit(0)
-  }
-  const wait = (delay) =>
-    new Promise((resolve) => {
-      setTimeout(resolve, delay)
-    })
-  const allPackages = pkgUtils.getPackages()
-  for (const pkg of allPackages) {
-    if (pkg.private) {
-      info(`${pkg.name} is private.`)
-    } else {
-      if (command === 'add') {
-        const toTag = `${pkg.name}@${versionToTag}`
-        info(
-          `📦 Running 'dist-tag ${command} ${toTag} ${tag}' for ${pkg.name}...`
-        )
-        try {
-          await runCommandAsync('npm', ['dist-tag', command, toTag, tag])
-        } catch (err) {
-          error(err)
-        }
-      } else if (command === 'rm') {
-        info(`📦 Running 'dist-tag ${command} ${tag}' for ${pkg.name}...`)
-        try {
-          await runCommandAsync('npm', ['dist-tag', command, tag])
-        } catch (err) {
-          error(err)
-        }
-      } else if (command === 'ls') {
-        info(`📦 Running 'dist-tag ${command}' for ${pkg.name} ...`)
-        try {
-          await runCommandAsync('npm', ['dist-tag', command])
-        } catch (err) {
-          error(err)
-        }
-      }
-      await wait(500)
+
+  try {
+    if (command === 'add') {
+      info(
+        `Command: "pnpm dist-tag ${command}". Version to tag as "${tag}": ${versionToTag}`
+      )
+    } else if (command === 'rm') {
+      info(`Command: "pnpm dist-tag ${command} ${tag}".`)
+    } else if (command === 'ls') {
+      info(`Command: "pnpm dist-tag ${command}".`)
     }
+    const reply = await confirm('Continue? [y/n]\n')
+    if (!['Y', 'y'].includes(reply.trim())) {
+      process.exit(0)
+    }
+    const wait = (delay) =>
+      new Promise((resolve) => {
+        setTimeout(resolve, delay)
+      })
+    const allPackages = pkgUtils.getPackages()
+    for (const pkg of allPackages) {
+      if (pkg.private) {
+        info(`${pkg.name} is private.`)
+      } else {
+        if (command === 'add') {
+          const toTag = `${pkg.name}@${versionToTag}`
+          info(
+            `📦 Running 'dist-tag ${command} ${toTag} ${tag}' for ${pkg.name}...`
+          )
+          try {
+            await runCommandAsync('pnpm', ['dist-tag', command, toTag, tag])
+          } catch (err) {
+            error(err)
+          }
+        } else if (command === 'rm') {
+          info(`📦 Running 'dist-tag ${command} ${tag}' for ${pkg.name}...`)
+          try {
+            await runCommandAsync('pnpm', ['dist-tag', command, tag])
+          } catch (err) {
+            error(err)
+          }
+        } else if (command === 'ls') {
+          info(`📦 Running 'dist-tag ${command}' for ${pkg.name} ...`)
+          try {
+            await runCommandAsync('pnpm', ['dist-tag', command])
+          } catch (err) {
+            error(err)
+          }
+        }
+        await wait(500)
+      }
+    }
+  } finally {
+    cleanupNPMRCFile()
   }
 }