+++ packages/ui-drilldown/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [11.2.0](https://github.com/instructure/instructure-ui/compare/v11.0.1...v11.2.0) (2025-11-06)
+
+### Features
+
+- **many:** migrate from npm to pnpm ([f7bb16e](https://github.com/instructure/instructure-ui/commit/f7bb16e114df83984c67d5a6e07fb4d9c65efc53))
+- **ui-drilldown:** make Drilldown.Group controlled via selectedOptions prop. If provided, this prop fully drives the selection state instead of internal state ([3341639](https://github.com/instructure/instructure-ui/commit/33416395bf7d9a4d5649334d65cc9b599dbc2ec3))
+
+# [11.1.0](https://github.com/instructure/instructure-ui/compare/v11.0.1...v11.1.0) (2025-11-05)
+
+### Features
+
+- **many:** migrate from npm to pnpm ([f7bb16e](https://github.com/instructure/instructure-ui/commit/f7bb16e114df83984c67d5a6e07fb4d9c65efc53))
+- **ui-drilldown:** make Drilldown.Group controlled via selectedOptions prop. If provided, this prop fully drives the selection state instead of internal state ([3341639](https://github.com/instructure/instructure-ui/commit/33416395bf7d9a4d5649334d65cc9b599dbc2ec3))
+
 ## [11.0.1](https://github.com/instructure/instructure-ui/compare/v11.0.0...v11.0.1) (2025-10-13)
 
 **Note:** Version bump only for package @instructure/ui-drilldown